Revert "fix microsecond issue on roundtrip tests in python3"
This reverts commit 291d712a92525480e8bfcec1eb1cc65331ad9917.

This issue was resolved in Python 3.4 and later in September 2015
(https://bugs.python.org/issue23517). This hack should not have been necessary
in 2017 when it was committed if using a recent version of Python, but certainly
is not necessary anymore.

The hack does not correctly handle the case where ms == 10000000. In that case,
setting millisecond=ms will crash since that value is out of the acceptable
range. In that case, we have to either add +1 second if Python has floored the
value (prior to the 3.4 fix above), or not add anything if Python has rounded.

Since <3.4 is extremely old and unlikely to be used anymore for production code,
the simplest course of action is to remove the hack.