update user leap data cache at module load if leap data invalid
this is somewhat dubious, but seems like the simplest way for users to get
the needed data if it's not available on their system.

Upon module load, the validity of the available leap data is checked.  if
it's invalid (unavailable or expired) the latest leap data is download from
the IETF and stored in a local user cache.  This should only happen if the
data is unavailable (so first usage for mac or windows user), or expired
(which happens only once every couple of years).